Populate `futures::{task, executor}` on `no_std`
This commit enables usage of the standard task system in `no_std` contexts. This
essentially piggy-backs the `use_std` feature already have and, when
deactivated, will provide the `futures::{task, executor}` modules. Building on
all the previous refactorings for types like `NotifyHandle` this should enable
usage of the task system in these contexts.

This does, however, come with a caveat. A relatively large dependency of the
task system is currently its usage of thread-local storage, unavailable in task
contexts! To work around this an initialization hook for the futures library is
provided, and this initialization hook is required to be called before working
with the task system. This hook takes just two pointers, a `get` and a `set` for
a pointer read/write into some "local storage", depending on the context of the
system that we're running on.

Usage of this crate from a `std` context (e.g. with default features enabled)
should not be impacted by this change. Everything previously present is still
there and no explicit initialization is required.
