+++ src/lib.rs
@@ -198,12 +198,9 @@ pub use future::{
 
 #[cfg(feature = "use_std")]
 mod lock;
-#[cfg(feature = "use_std")]
 mod task_impl;
 
-#[cfg(feature = "use_std")]
 pub mod task;
-#[cfg(feature = "use_std")]
 pub mod executor;
 #[cfg(feature = "use_std")]
 pub mod sync;
