+++ src/task.rs
@@ -29,17 +29,18 @@
 
 #[doc(hidden)]
 #[deprecated(since = "0.1.4", note = "import through the executor module instead")]
-#[cfg(feature = "with-deprecated")]
+#[cfg(all(feature = "with-deprecated", feature = "use_std"))]
 #[allow(deprecated)]
-pub use task_impl::{Spawn, spawn, Unpark, Executor, Run};
+pub use task_impl::{Spawn, spawn, Unpark, Executor, Run, park};
+
+pub use task_impl::{Task, current, init};
 
 #[allow(deprecated)]
-pub use task_impl::{Task, park, current};
-#[allow(deprecated)]
+#[cfg(feature = "use_std")]
 pub use task_impl::{LocalKey, with_unpark_event, UnparkEvent, EventSet};
 
 #[doc(hidden)]
 #[deprecated(since = "0.1.4", note = "import through the executor module instead")]
-#[cfg(feature = "with-deprecated")]
+#[cfg(all(feature = "with-deprecated", feature = "use_std"))]
 #[allow(deprecated)]
 pub use task_impl::TaskRc;
