+++ src/task_impl/mod.rs
@@ -1,55 +1,30 @@
-use std::prelude::v1::*;
+use core::fmt;
+use core::marker::PhantomData;
 
-use std::cell::Cell;
-use std::fmt;
-use std::marker::PhantomData;
-use std::mem;
-use std::ptr;
-use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::thread;
-
-use {Poll, Future, Async, Stream, Sink, StartSend, AsyncSink};
-use future::BoxFuture;
-
-mod unpark_mutex;
-use self::unpark_mutex::UnparkMutex;
-
-mod data;
-pub use self::data::LocalKey;
+use {Poll, Future, Stream, Sink, StartSend};
 
 mod atomic_task;
 pub use self::atomic_task::AtomicTask;
 
-mod task_rc;
-#[allow(deprecated)]
-#[cfg(feature = "with-deprecated")]
-pub use self::task_rc::TaskRc;
+mod core;
+
+#[cfg(feature = "use_std")]
+mod std;
+#[cfg(feature = "use_std")]
+pub use self::std::*;
+#[cfg(not(feature = "use_std"))]
+pub use self::core::*;
 
 struct BorrowedTask<'a> {
     id: usize,
     unpark: BorrowedUnpark<'a>,
     events: BorrowedEvents<'a>,
     // Task-local storage
-    map: &'a data::LocalMap,
-}
-
-#[derive(Copy, Clone)]
-#[allow(deprecated)]
-enum BorrowedUnpark<'a> {
-    Old(&'a Arc<Unpark>),
-    New(&'a Fn() -> NotifyHandle, u64),
-}
-
-#[derive(Copy, Clone)]
-#[allow(deprecated)]
-enum BorrowedEvents<'a> {
-    None,
-    One(&'a UnparkEvent, &'a BorrowedEvents<'a>),
+    map: &'a LocalMap,
 }
 
 fn fresh_task_id() -> usize {
-    use std::sync::atomic::{AtomicUsize, Ordering, ATOMIC_USIZE_INIT};
+    use core::sync::atomic::{AtomicUsize, Ordering, ATOMIC_USIZE_INIT};
 
     // TODO: this assert is a real bummer, need to figure out how to reuse
     //       old IDs that are no longer in use.
@@ -63,36 +38,26 @@ fn fresh_task_id() -> usize {
     id
 }
 
-thread_local!(static CURRENT_TASK: Cell<*const BorrowedTask<'static>> = {
-    Cell::new(0 as *const _)
-});
-
 fn set<'a, F, R>(task: &BorrowedTask<'a>, f: F) -> R
     where F: FnOnce() -> R
 {
-    struct Reset(*const BorrowedTask<'static>);
+    struct Reset(*mut u8);
     impl Drop for Reset {
         fn drop(&mut self) {
-            CURRENT_TASK.with(|c| c.set(self.0));
+            set_ptr(self.0);
         }
     }
 
-    CURRENT_TASK.with(move |c| {
-        let _reset = Reset(c.get());
-        let task = unsafe {
-            mem::transmute::<&BorrowedTask<'a>,
-                             *const BorrowedTask<'static>>(task)
-        };
-        c.set(task);
-        f()
-    })
+    let _reset = Reset(get_ptr());
+    set_ptr(task as *const BorrowedTask<'a> as *mut u8);
+    f()
 }
 
 fn with<F: FnOnce(&BorrowedTask) -> R, R>(f: F) -> R {
-    let task = CURRENT_TASK.with(|c| c.get());
-    assert!(!task.is_null(), "no Task is currently running");
     unsafe {
-        f(&*task)
+        let task = get_ptr();
+        assert!(!task.is_null(), "no Task is currently running");
+        f(&*(task as *const BorrowedTask))
     }
 }
 
@@ -104,6 +69,7 @@ fn with<F: FnOnce(&BorrowedTask) -> R, R>(f: F) -> R {
 /// the future as notifications arrive, until the future terminates.
 ///
 /// This is obtained by the `task::current` function.
+#[derive(Clone)]
 pub struct Task {
     id: usize,
     unpark: TaskUnpark,
@@ -113,20 +79,6 @@ pub struct Task {
 fn _assert_kinds() {
     fn _assert_send<T: Send>() {}
     _assert_send::<Task>();
-}
-
-enum TaskUnpark {
-    #[allow(deprecated)]
-    Old(Arc<Unpark>),
-    New(NotifyHandle, u64),
-}
-
-#[derive(Clone)]
-#[allow(deprecated)]
-enum UnparkEvents {
-    None,
-    One(UnparkEvent),
-    Many(Box<[UnparkEvent]>),
 }
 
 /// Returns a handle to the current task to call `notify` at a later date.
@@ -153,37 +105,8 @@ enum UnparkEvents {
 /// `poll`.
 pub fn current() -> Task {
     with(|borrowed| {
-        let unpark = match borrowed.unpark {
-            BorrowedUnpark::Old(old) => TaskUnpark::Old(old.clone()),
-            BorrowedUnpark::New(new, id) => {
-                // A new handle is being created, increment the ref count
-                let handle = new();
-                let id = handle.clone_id(id);
-                TaskUnpark::New(handle, id)
-            }
-        };
-
-        let mut one_event = None;
-        let mut list = Vec::new();
-        let mut cur = &borrowed.events;
-        while let BorrowedEvents::One(event, next) = *cur {
-            let event = event.clone();
-            match one_event.take() {
-                None if list.len() == 0 => one_event = Some(event),
-                None => list.push(event),
-                Some(event2) =>  {
-                    list.push(event2);
-                    list.push(event);
-                }
-            }
-            cur = next;
-        }
-
-        let events = match one_event {
-            None if list.len() == 0 => UnparkEvents::None,
-            None => UnparkEvents::Many(list.into_boxed_slice()),
-            Some(e) => UnparkEvents::One(e),
-        };
+        let unpark = borrowed.unpark.to_owned();
+        let events = borrowed.events.to_owned();
 
         Task {
             id: borrowed.id,
@@ -208,22 +131,9 @@ impl Task {
     /// If the task is currently polling its future when `notify` is called, it
     /// must poll the future *again* afterwards, ensuring that all relevant
     /// events are eventually observed by the future.
-    #[allow(deprecated)]
     pub fn notify(&self) {
-        match self.events {
-            UnparkEvents::None => {}
-            UnparkEvents::One(ref e) => e.unpark(),
-            UnparkEvents::Many(ref list) => {
-                for event in list.iter() {
-                    event.unpark();
-                }
-            }
-        }
-
-        match self.unpark {
-            TaskUnpark::Old(ref old) => old.unpark(),
-            TaskUnpark::New(ref new, id) => new.notify(id),
-        }
+        self.events.notify();
+        self.unpark.notify();
     }
 
     #[doc(hidden)]
@@ -294,35 +204,8 @@ impl Task {
     #[allow(deprecated)]
     pub fn will_notify_current(&self) -> bool {
         with(|current| {
-            let unpark_same = match (&current.unpark, &self.unpark) {
-                (&BorrowedUnpark::Old(old1),
-                 &TaskUnpark::Old(ref old2)) => {
-                    &**old1 as *const Unpark == &**old2 as *const Unpark
-                }
-                (&BorrowedUnpark::New(new1, id1),
-                 &TaskUnpark::New(ref new2, id2)) => {
-                    // TODO: should cache this invocation of `new1` in the
-                    //       borrowed storage so we know we have an owned value
-                    //       on hand.
-                    id1 == id2 && new1().inner == new2.inner
-                }
-                _ => false,
-            };
-            if !unpark_same {
-                return false
-            }
-
-            // Pessimistically assume that any unpark events mean that we're not
-            // equivalent to the current task.
-            match self.events {
-                UnparkEvents::None => {}
-                _ => return false,
-            }
-
-            match current.events {
-                BorrowedEvents::None => true,
-                _ => false,
-            }
+            self.unpark.will_notify(&current.unpark) &&
+                self.events.will_notify(&current.events)
         })
     }
 }
@@ -331,32 +214,6 @@ impl fmt::Debug for Task {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_struct("Task")
          .finish()
-    }
-}
-
-impl Clone for Task {
-    fn clone(&self) -> Task {
-        let unpark = match self.unpark {
-            TaskUnpark::Old(ref old) => TaskUnpark::Old(old.clone()),
-            TaskUnpark::New(ref new, id) => {
-                let id = new.clone_id(id);
-                TaskUnpark::New(new.clone(), id)
-            }
-        };
-
-        Task {
-            id: self.id,
-            unpark: unpark,
-            events: self.events.clone(),
-        }
-    }
-}
-
-impl Drop for Task {
-    fn drop(&mut self) {
-        if let TaskUnpark::New(ref new, id) = self.unpark {
-            new.drop_id(id);
-        }
     }
 }
 
@@ -373,7 +230,7 @@ impl Drop for Task {
 pub struct Spawn<T> {
     id: usize,
     obj: T,
-    data: data::LocalMap,
+    data: LocalMap,
 }
 
 /// Spawns a new future, returning the fused future and task.
@@ -389,7 +246,7 @@ pub fn spawn<T>(obj: T) -> Spawn<T> {
     Spawn {
         id: fresh_task_id(),
         obj: obj,
-        data: data::local_map(),
+        data: local_map(),
     }
 }
 
@@ -411,25 +268,6 @@ impl<T> Spawn<T> {
 }
 
 impl<F: Future> Spawn<F> {
-    /// Polls the internal future, scheduling notifications to be sent to the
-    /// `unpark` argument.
-    ///
-    /// This method will poll the internal future, testing if it's completed
-    /// yet. The `unpark` argument is used as a sink for notifications sent to
-    /// this future. That is, while the future is being polled, any call to
-    /// `task::park()` will return a handle that contains the `unpark`
-    /// specified.
-    ///
-    /// If this function returns `NotReady`, then the `unpark` should have been
-    /// scheduled to receive a notification when poll can be called again.
-    /// Otherwise if `Ready` or `Err` is returned, the `Spawn` task can be
-    /// safely destroyed.
-    #[deprecated(note = "recommended to use `poll_future_notify` instead")]
-    #[allow(deprecated)]
-    pub fn poll_future(&mut self, unpark: Arc<Unpark>) -> Poll<F::Item, F::Error> {
-        self.enter(BorrowedUnpark::Old(&unpark), |f| f.poll())
-    }
-
     /// Polls the internal future, scheduling notifications to be sent to the
     /// `notify` argument.
     ///
@@ -465,67 +303,11 @@ impl<F: Future> Spawn<F> {
         where T: Clone + Into<NotifyHandle>,
     {
         let mk = || notify.clone().into();
-        self.enter(BorrowedUnpark::New(&mk, id), |f| f.poll())
-    }
-
-    /// Waits for the internal future to complete, blocking this thread's
-    /// execution until it does.
-    ///
-    /// This function will call `poll_future` in a loop, waiting for the future
-    /// to complete. When a future cannot make progress it will use
-    /// `thread::park` to block the current thread.
-    pub fn wait_future(&mut self) -> Result<F::Item, F::Error> {
-        let unpark = Arc::new(ThreadUnpark::new(thread::current()));
-
-        loop {
-            match try!(self.poll_future_notify(&unpark, 0)) {
-                Async::NotReady => unpark.park(),
-                Async::Ready(e) => return Ok(e),
-            }
-        }
-    }
-
-    /// A specialized function to request running a future to completion on the
-    /// specified executor.
-    ///
-    /// This function only works for futures whose item and error types are `()`
-    /// and also implement the `Send` and `'static` bounds. This will submit
-    /// units of work (instances of `Run`) to the `exec` argument provided
-    /// necessary to drive the future to completion.
-    ///
-    /// When the future would block, it's arranged that when the future is again
-    /// ready it will submit another unit of work to the `exec` provided. This
-    /// will happen in a loop until the future has completed.
-    ///
-    /// This method is not appropriate for all futures, and other kinds of
-    /// executors typically provide a similar function with perhaps relaxed
-    /// bounds as well.
-    pub fn execute(self, exec: Arc<Executor>)
-        where F: Future<Item=(), Error=()> + Send + 'static,
-    {
-        exec.clone().execute(Run {
-            // Ideally this method would be defined directly on
-            // `Spawn<BoxFuture<(), ()>>` so we wouldn't have to box here and
-            // it'd be more explicit, but unfortunately that currently has a
-            // link error on nightly: rust-lang/rust#36155
-            spawn: spawn(self.into_inner().boxed()),
-            inner: Arc::new(RunInner {
-                exec: exec,
-                mutex: UnparkMutex::new()
-            }),
-        })
+        self.enter(BorrowedUnpark::new(&mk, id), |f| f.poll())
     }
 }
 
 impl<S: Stream> Spawn<S> {
-    /// Like `poll_future`, except polls the underlying stream.
-    #[deprecated(note = "recommended to use `poll_stream_notify` instead")]
-    #[allow(deprecated)]
-    pub fn poll_stream(&mut self, unpark: Arc<Unpark>)
-                       -> Poll<Option<S::Item>, S::Error> {
-        self.enter(BorrowedUnpark::Old(&unpark), |s| s.poll())
-    }
-
     /// Like `poll_future_notify`, except polls the underlying stream.
     pub fn poll_stream_notify<T>(&mut self,
                                  notify: &T,
@@ -534,37 +316,11 @@ impl<S: Stream> Spawn<S> {
         where T: Clone + Into<NotifyHandle>,
     {
         let mk = || notify.clone().into();
-        self.enter(BorrowedUnpark::New(&mk, id), |s| s.poll())
-    }
-
-    /// Like `wait_future`, except only waits for the next element to arrive on
-    /// the underlying stream.
-    pub fn wait_stream(&mut self) -> Option<Result<S::Item, S::Error>> {
-        let unpark = Arc::new(ThreadUnpark::new(thread::current()));
-        loop {
-            match self.poll_stream_notify(&unpark, 0) {
-                Ok(Async::NotReady) => unpark.park(),
-                Ok(Async::Ready(Some(e))) => return Some(Ok(e)),
-                Ok(Async::Ready(None)) => return None,
-                Err(e) => return Some(Err(e)),
-            }
-        }
+        self.enter(BorrowedUnpark::new(&mk, id), |s| s.poll())
     }
 }
 
 impl<S: Sink> Spawn<S> {
-    /// Invokes the underlying `start_send` method with this task in place.
-    ///
-    /// If the underlying operation returns `NotReady` then the `unpark` value
-    /// passed in will receive a notification when the operation is ready to be
-    /// attempted again.
-    #[deprecated(note = "recommended to use `start_send_notify` instead")]
-    #[allow(deprecated)]
-    pub fn start_send(&mut self, value: S::SinkItem, unpark: &Arc<Unpark>)
-                       -> StartSend<S::SinkItem, S::SinkError> {
-        self.enter(BorrowedUnpark::Old(unpark), |s| s.start_send(value))
-    }
-
     /// Invokes the underlying `start_send` method with this task in place.
     ///
     /// If the underlying operation returns `NotReady` then the `notify` value
@@ -578,19 +334,7 @@ impl<S: Sink> Spawn<S> {
         where T: Clone + Into<NotifyHandle>,
     {
         let mk = || notify.clone().into();
-        self.enter(BorrowedUnpark::New(&mk, id), |s| s.start_send(value))
-    }
-
-    /// Invokes the underlying `poll_complete` method with this task in place.
-    ///
-    /// If the underlying operation returns `NotReady` then the `unpark` value
-    /// passed in will receive a notification when the operation is ready to be
-    /// attempted again.
-    #[deprecated(note = "recommended to use `poll_flush_notify` instead")]
-    #[allow(deprecated)]
-    pub fn poll_flush(&mut self, unpark: &Arc<Unpark>)
-                       -> Poll<(), S::SinkError> {
-        self.enter(BorrowedUnpark::Old(unpark), |s| s.poll_complete())
+        self.enter(BorrowedUnpark::new(&mk, id), |s| s.start_send(value))
     }
 
     /// Invokes the underlying `poll_complete` method with this task in place.
@@ -605,42 +349,7 @@ impl<S: Sink> Spawn<S> {
         where T: Clone + Into<NotifyHandle>,
     {
         let mk = || notify.clone().into();
-        self.enter(BorrowedUnpark::New(&mk, id), |s| s.poll_complete())
-    }
-
-    /// Blocks the current thread until it's able to send `value` on this sink.
-    ///
-    /// This function will send the `value` on the sink that this task wraps. If
-    /// the sink is not ready to send the value yet then the current thread will
-    /// be blocked until it's able to send the value.
-    pub fn wait_send(&mut self, mut value: S::SinkItem)
-                     -> Result<(), S::SinkError> {
-        let notify = Arc::new(ThreadUnpark::new(thread::current()));
-        loop {
-            value = match try!(self.start_send_notify(value, &notify, 0)) {
-                AsyncSink::NotReady(v) => v,
-                AsyncSink::Ready => return Ok(()),
-            };
-            notify.park();
-        }
-    }
-
-    /// Blocks the current thread until it's able to flush this sink.
-    ///
-    /// This function will call the underlying sink's `poll_complete` method
-    /// until it returns that it's ready, proxying out errors upwards to the
-    /// caller if one occurs.
-    ///
-    /// The thread will be blocked until `poll_complete` returns that it's
-    /// ready.
-    pub fn wait_flush(&mut self) -> Result<(), S::SinkError> {
-        let notify = Arc::new(ThreadUnpark::new(thread::current()));
-        loop {
-            if try!(self.poll_flush_notify(&notify, 0)).is_ready() {
-                return Ok(())
-            }
-            notify.park();
-        }
+        self.enter(BorrowedUnpark::new(&mk, id), |s| s.poll_complete())
     }
 }
 
@@ -651,7 +360,7 @@ impl<T> Spawn<T> {
         let borrowed = BorrowedTask {
             id: self.id,
             unpark: unpark,
-            events: BorrowedEvents::None,
+            events: BorrowedEvents::new(),
             map: &self.data,
         };
         let obj = &mut self.obj;
@@ -664,92 +373,6 @@ impl<T: fmt::Debug> fmt::Debug for Spawn<T> {
         f.debug_struct("Spawn")
          .field("obj", &self.obj)
          .finish()
-    }
-}
-
-/// A trait which represents a sink of notifications that a future is ready to
-/// make progress.
-///
-/// This trait is provided as an argument to the `Spawn::poll_future` and
-/// `Spawn::poll_stream` functions. It's transitively used as part of the
-/// `Task::unpark` method to internally deliver notifications of readiness of a
-/// future to move forward.
-#[deprecated(note = "recommended to use `Notify` instead")]
-pub trait Unpark: Send + Sync {
-    /// Indicates that an associated future and/or task are ready to make
-    /// progress.
-    ///
-    /// Typically this means that the receiver of the notification should
-    /// arrange for the future to get poll'd in a prompt fashion.
-    fn unpark(&self);
-}
-
-/// A trait representing requests to poll futures.
-///
-/// This trait is an argument to the `Spawn::execute` which is used to run a
-/// future to completion. An executor will receive requests to run a future and
-/// an executor is responsible for ensuring that happens in a timely fashion.
-pub trait Executor: Send + Sync + 'static {
-    /// Requests that `Run` is executed soon on the given executor.
-    fn execute(&self, r: Run);
-}
-
-/// Units of work submitted to an `Executor`, currently only created
-/// internally.
-pub struct Run {
-    spawn: Spawn<BoxFuture<(), ()>>,
-    inner: Arc<RunInner>,
-}
-
-struct RunInner {
-    mutex: UnparkMutex<Run>,
-    exec: Arc<Executor>,
-}
-
-impl Run {
-    /// Actually run the task (invoking `poll` on its future) on the current
-    /// thread.
-    #[allow(deprecated)]
-    pub fn run(self) {
-        let Run { mut spawn, inner } = self;
-
-        // SAFETY: the ownership of this `Run` object is evidence that
-        // we are in the `POLLING`/`REPOLL` state for the mutex.
-        unsafe {
-            inner.mutex.start_poll();
-
-            loop {
-                let unpark: Arc<Unpark> = inner.clone();
-                match spawn.poll_future(unpark) {
-                    Ok(Async::NotReady) => {}
-                    Ok(Async::Ready(())) |
-                    Err(()) => return inner.mutex.complete(),
-                }
-                let run = Run { spawn: spawn, inner: inner.clone() };
-                match inner.mutex.wait(run) {
-                    Ok(()) => return,            // we've waited
-                    Err(r) => spawn = r.spawn,   // someone's notified us
-                }
-            }
-        }
-    }
-}
-
-impl fmt::Debug for Run {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("Run")
-         .field("contents", &"...")
-         .finish()
-    }
-}
-
-#[allow(deprecated)]
-impl Unpark for RunInner {
-    fn unpark(&self) {
-        match self.mutex.notify() {
-            Ok(run) => self.exec.execute(run),
-            Err(()) => {}
-        }
     }
 }
 
@@ -816,76 +439,6 @@ pub trait Notify: Send + Sync {
     }
 }
 
-// ===== ThreadUnpark =====
-
-struct ThreadUnpark {
-    thread: thread::Thread,
-    ready: AtomicBool,
-}
-
-impl ThreadUnpark {
-    fn new(thread: thread::Thread) -> ThreadUnpark {
-        ThreadUnpark {
-            thread: thread,
-            ready: AtomicBool::new(false),
-        }
-    }
-
-    fn park(&self) {
-        if !self.ready.swap(false, Ordering::SeqCst) {
-            thread::park();
-        }
-    }
-}
-
-impl Notify for ThreadUnpark {
-    fn notify(&self, _unpark_id: u64) {
-        self.ready.store(true, Ordering::SeqCst);
-        self.thread.unpark()
-    }
-}
-
-// ===== UnparkEvent =====
-
-/// For the duration of the given callback, add an "unpark event" to be
-/// triggered when the task handle is used to unpark the task.
-///
-/// Unpark events are used to pass information about what event caused a task to
-/// be unparked. In some cases, tasks are waiting on a large number of possible
-/// events, and need precise information about the wakeup to avoid extraneous
-/// polling.
-///
-/// Every `Task` handle comes with a set of unpark events which will fire when
-/// `unpark` is called. When fired, these events insert an identifer into a
-/// concurrent set, which the task can read from to determine what events
-/// occurred.
-///
-/// This function immediately invokes the closure, `f`, but arranges things so
-/// that `task::park` will produce a `Task` handle that includes the given
-/// unpark event.
-///
-/// # Panics
-///
-/// This function will panic if a task is not currently being executed. That
-/// is, this method can be dangerous to call outside of an implementation of
-/// `poll`.
-#[deprecated(note = "recommended to use `FuturesUnordered` instead")]
-#[allow(deprecated)]
-pub fn with_unpark_event<F, R>(event: UnparkEvent, f: F) -> R
-    where F: FnOnce() -> R
-{
-    with(|task| {
-        let new_task = BorrowedTask {
-            id: task.id,
-            unpark: task.unpark,
-            events: BorrowedEvents::One(&event, &task.events),
-            map: task.map,
-        };
-
-        set(&new_task, f)
-    })
-}
-
 pub fn with_notify<F, T, R>(notify: &T, id: u64, f: F) -> R
     where F: FnOnce() -> R,
           T: Clone + Into<NotifyHandle>,
@@ -894,63 +447,13 @@ pub fn with_notify<F, T, R>(notify: &T, id: u64, f: F) -> R
         let mk = || notify.clone().into();
         let new_task = BorrowedTask {
             id: task.id,
-            unpark: BorrowedUnpark::New(&mk, id),
+            unpark: BorrowedUnpark::new(&mk, id),
             events: task.events,
             map: task.map,
         };
 
         set(&new_task, f)
     })
-}
-
-/// A set insertion to trigger upon `unpark`.
-///
-/// Unpark events are used to communicate information about *why* an unpark
-/// occured, in particular populating sets with event identifiers so that the
-/// unparked task can avoid extraneous polling. See `with_unpark_event` for
-/// more.
-#[derive(Clone)]
-#[deprecated(note = "recommended to use `FuturesUnordered` instead")]
-#[allow(deprecated)]
-pub struct UnparkEvent {
-    set: Arc<EventSet>,
-    item: usize,
-}
-
-#[allow(deprecated)]
-impl UnparkEvent {
-    /// Construct an unpark event that will insert `id` into `set` when
-    /// triggered.
-    #[deprecated(note = "recommended to use `FuturesUnordered` instead")]
-    pub fn new(set: Arc<EventSet>, id: usize) -> UnparkEvent {
-        UnparkEvent {
-            set: set,
-            item: id,
-        }
-    }
-
-    fn unpark(&self) {
-        self.set.insert(self.item);
-    }
-}
-
-#[allow(deprecated)]
-impl fmt::Debug for UnparkEvent {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("UnparkEvent")
-         .field("set", &"...")
-         .field("item", &self.item)
-         .finish()
-    }
-}
-
-/// A concurrent set which allows for the insertion of `usize` values.
-///
-/// `EventSet`s are used to communicate precise information about the event(s)
-/// that trigged a task notification. See `task::with_unpark_event` for details.
-pub trait EventSet: Send + Sync + 'static {
-    /// Insert the given ID into the set
-    fn insert(&self, id: usize);
 }
 
 /// An unsafe trait for implementing custom forms of memory management behind a
@@ -1124,76 +627,6 @@ impl Drop for NotifyHandle {
     }
 }
 
-// Safe implementation of `UnsafeNotify` for `Arc` in the standard library.
-//
-// Note that this is a very unsafe implementation! The crucial pieces is that
-// these two values are considered equivalent:
-//
-// * Arc<T>
-// * *const ArcWrapped<T>
-//
-// We don't actually know the layout of `ArcWrapped<T>` as it's an
-// implementation detail in the standard library. We can work, though, by
-// casting it through and back an `Arc<T>`.
-//
-// This also means that you wn't actually fine `UnsafeNotify for Arc<T>`
-// because it's the wrong level of indirection. These methods are sort of
-// receiving Arc<T>, but not an owned version. It's... complicated. We may be
-// one of the first users of unsafe trait objects!
-
-struct ArcWrapped<T>(PhantomData<T>);
-
-impl<T: Notify + 'static> Notify for ArcWrapped<T> {
-    fn notify(&self, id: u64) {
-        unsafe {
-            let me: *const ArcWrapped<T> = self;
-            T::notify(&*(&me as *const *const ArcWrapped<T> as *const Arc<T>),
-                      id)
-        }
-    }
-
-    fn clone_id(&self, id: u64) -> u64 {
-        unsafe {
-            let me: *const ArcWrapped<T> = self;
-            T::clone_id(&*(&me as *const *const ArcWrapped<T> as *const Arc<T>),
-                        id)
-        }
-    }
-
-    fn drop_id(&self, id: u64) {
-        unsafe {
-            let me: *const ArcWrapped<T> = self;
-            T::drop_id(&*(&me as *const *const ArcWrapped<T> as *const Arc<T>),
-                       id)
-        }
-    }
-}
-
-unsafe impl<T: Notify + 'static> UnsafeNotify for ArcWrapped<T> {
-    unsafe fn clone_raw(&self) -> NotifyHandle {
-        let me: *const ArcWrapped<T> = self;
-        let arc = (*(&me as *const *const ArcWrapped<T> as *const Arc<T>)).clone();
-        NotifyHandle::from(arc)
-    }
-
-    unsafe fn drop_raw(&self) {
-        let mut me: *const ArcWrapped<T> = self;
-        let me = &mut me as *mut *const ArcWrapped<T> as *mut Arc<T>;
-        ptr::drop_in_place(me);
-    }
-}
-
-impl<T> From<Arc<T>> for NotifyHandle
-    where T: Notify + 'static,
-{
-    fn from(rc: Arc<T>) -> NotifyHandle {
-        unsafe {
-            let ptr = mem::transmute::<Arc<T>, *mut ArcWrapped<T>>(rc);
-            NotifyHandle::new(ptr)
-        }
-    }
-}
-
 /// Marker for a `T` that is behind &'static.
 struct StaticRef<T>(PhantomData<T>);
 
@@ -1227,4 +660,3 @@ impl<T: Notify> From<&'static T> for NotifyHandle {
         unsafe { NotifyHandle::new(src as *const _ as *mut StaticRef<T>) }
     }
 }
-
