+++ src/task_impl/std/data.rs
@@ -5,6 +5,8 @@ use std::cell::RefCell;
 use std::hash::{BuildHasherDefault, Hasher};
 use std::collections::HashMap;
 
+use task_impl::with;
+
 /// A macro to create a `static` of type `LocalKey`
 ///
 /// This macro is intentionally similar to the `thread_local!`, and creates a
@@ -113,7 +115,7 @@ impl<T: Send + 'static> LocalKey<T> {
         where F: FnOnce(&T) -> R
     {
         let key = (self.__key)();
-        super::with(|task| {
+        with(|task| {
             let raw_pointer = {
                 let mut data = task.map.borrow_mut();
                 let entry = data.entry(key).or_insert_with(|| {
